# docs/swagger.yaml
definitions:
  models.Project:
    properties:
      description:
        maxLength: 500
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/models.User'
        type: array
      name:
        maxLength: 10
        minLength: 3
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    required:
    - name
    type: object
  models.Subtask:
    properties:
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creator_id:
        type: string
      id:
        type: string
      status:
        type: string
      task:
        $ref: '#/definitions/models.Task'
      task_id:
        type: string
      title:
        minLength: 3
        type: string
    required:
    - status
    - task_id
    - title
    type: object
  models.Task:
    properties:
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creator_id:
        type: string
      description:
        maxLength: 500
        type: string
      id:
        type: string
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      status:
        enum:
        - pending
        - in_progress
        - completed
        type: string
      subtasks:
        items:
          $ref: '#/definitions/models.Subtask'
        type: array
      title:
        type: string
    required:
    - project_id
    - status
    - title
    type: object
  models.User:
    properties:
      id:
        type: string
      password:
        minLength: 6
        type: string
      projects:
        items:
          $ref: '#/definitions/models.Project'
        type: array
      role:
        enum:
        - admin
        - manager
        - team_member
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - role
    - username
    type: object
  utils.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
  /admin/users:
    get:
      description: Get all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
  /admin/users/{id}:
    delete:
      description: Delete a user (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
...
  /createtask:
    post:
      consumes:
      - application/json
      description: Create a task in a project
      parameters:
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a task
      tags:
      - Tasks
...
  /gettask:
    get:
      description: Get tasks accessible to the user with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List tasks
      tags:
      - Tasks
...
  /tasks/{id}:
    delete:
      description: Delete a task if the user is the creator or admin
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update a task if the user is the creator, assignee, or admin
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Tasks
  /users/{id}:
    put:
      consumes:
      - application/json
      description: Update a user's details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
swagger: "2.0"